refactor: put reused function into utils
